requirements.txt
# 成交魔方 V3 - 依赖管理
# 更新时间: 2026-03-30

# ==================== Web框架 ====================
streamlit>=1.32.0
streamlit-option-menu>=0.3.6

# ==================== 数据库 ====================
supabase>=2.0.3
postgrest>=0.16.3

# ==================== AI服务 ====================
openai>=1.0.0

# ==================== 配置管理 ====================
python-dotenv>=1.0.0

# ==================== 数据验证 ====================
pydantic>=2.0.0
typing-extensions>=4.0.0

# ==================== 数据处理 ====================
pandas>=2.0.0
diskcache>=5.6.0

# ==================== 工具库 ====================
python-dateutil>=2.8.0
requests>=2.31.0
cachetools>=5.0.0
structlog>=23.0.0
